requirements.txt
python-dotenv==0.12.0
mysqlclient==1.4.6
mypy_extensions==0.4.3
jsonschema==3.2.0
flask==1.1.1
gunicorn==20.0.4
pycodestyle==2.6.0
gitchangelog==3.0.4
google-api-python-client==1.8.3
autopep8==1.5.2
isort==4.3.21
pyjwt==1.7.1